backend/requirements.txt
asgiref==3.5.2
astroid==2.12.13
attrs==22.2.0
autopep8==2.0.0
dill==0.3.6
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
packaging==23.0
platformdirs==2.5.4
pluggy==1.0.0
pycodestyle==2.10.0
pylint==2.15.8
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytest==7.2.1
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
wrapt==1.14.1
